streamlit_app/requirements.txt
streamlit>=1.28.0
web3>=6.0.0
pandas>=1.5.0
plotly>=5.15.0
python-dotenv>=1.0.0 